Extract usePanelState composable from EditorApp.vue
Move panel state management (~90 lines) into dedicated composable:
- 6 panel expanded refs + 5 visibility refs
- 2 auto-advance settings (MIDI learn, paste)
- Row index display preference (hex/decimal)
- All localStorage keys, initialization, and watch persistence

Zero functional changes - pure refactoring.
