+++ editor/src/components/EditorApp.vue
@@ -44,6 +44,7 @@ import { useActionHistory, type DeserializationContext } from '../composables/us
 import { useVariableUsage } from '../composables/useVariableUsage';
 import { useSkipAnalysis } from '../composables/useSkipAnalysis';
 import { useRowSelection } from '../composables/useRowSelection';
+import { usePanelState } from '../composables/usePanelState';
 import { SetRowColorCommand, SetRowCommentCommand } from '../commands';
 import { MIDI_LEARN_FUNCTION_KEY } from '../constants/midi';
 import { COLOR_PALETTE } from '../constants/colors';
@@ -95,95 +96,31 @@ function showToast(message: string, type: 'success' | 'info' | 'warning' | 'erro
 }
 
 
-// Settings Panel expanded state (for MIDI Monitor positioning)
-const settingsPanelExpanded = ref(false);
-
-// MIDI Monitor expanded state (for Variable Monitor positioning)
-const midiMonitorExpanded = ref(false);
-
-// Variable Monitor expanded state (for toolbar positioning)
-const variableMonitorExpanded = ref(false);
-
-// Selection Toolbar expanded state
-const selectionToolbarExpanded = ref(false);
-
-// Log Monitor expanded state
-const logMonitorExpanded = ref(false);
-
-// Global Documentation Panel expanded state
-const getInitialGlobalDocExpanded = (): boolean => {
-  const stored = localStorage.getItem('global-doc-panel-expanded');
-  return stored ? JSON.parse(stored) : false;
-};
-const globalDocPanelExpanded = ref(getInitialGlobalDocExpanded());
-
-// Persist expanded state changes
-watch(globalDocPanelExpanded, (newValue) => {
-  localStorage.setItem('global-doc-panel-expanded', JSON.stringify(newValue));
-});
-
-// Panel visibility state (controlled from Settings Panel)
-const SELECTION_TOOLBAR_VISIBLE_KEY = 'nerdseq-show-selection-toolbar';
-const MIDI_MONITOR_VISIBLE_KEY = 'nerdseq-show-midi-monitor';
-const VARIABLE_MONITOR_VISIBLE_KEY = 'nerdseq-show-variable-monitor';
-const LOG_MONITOR_VISIBLE_KEY = 'nerdseq-show-log-monitor';
-const DESCRIPTION_VISIBLE_KEY = 'nerdseq-show-description';
-const MIDI_LEARN_AUTO_ADVANCE_KEY = 'nerdseq-midi-learn-auto-advance';
-const PASTE_AUTO_ADVANCE_KEY = 'nerdseq-paste-auto-advance';
-// Selection Toolbar defaults to true (enabled by default)
-const showSelectionToolbar = ref(localStorage.getItem(SELECTION_TOOLBAR_VISIBLE_KEY) !== 'false');
-const showMidiMonitor = ref(localStorage.getItem(MIDI_MONITOR_VISIBLE_KEY) === 'true');
-const showVariableMonitor = ref(localStorage.getItem(VARIABLE_MONITOR_VISIBLE_KEY) === 'true');
-const showLogMonitor = ref(localStorage.getItem(LOG_MONITOR_VISIBLE_KEY) === 'true');
-const showDescription = ref(localStorage.getItem(DESCRIPTION_VISIBLE_KEY) === 'true');
-const midiLearnAutoAdvance = ref(localStorage.getItem(MIDI_LEARN_AUTO_ADVANCE_KEY) !== 'false'); // Default: true
-
-// Initialize paste auto-advance setting with proper type handling
-const storedPasteAutoAdvance = localStorage.getItem(PASTE_AUTO_ADVANCE_KEY);
-const pasteAutoAdvance = ref<'disabled' | 'rows-only' | 'all'>(
-  storedPasteAutoAdvance === 'disabled' || storedPasteAutoAdvance === 'rows-only' || storedPasteAutoAdvance === 'all'
-    ? storedPasteAutoAdvance
-    : 'all' // Default: all
-);
+// Panel state composable (expanded/visibility/localStorage persistence)
+const {
+  settingsPanelExpanded,
+  midiMonitorExpanded,
+  variableMonitorExpanded,
+  selectionToolbarExpanded,
+  logMonitorExpanded,
+  globalDocPanelExpanded,
+  showSelectionToolbar,
+  showMidiMonitor,
+  showVariableMonitor,
+  showLogMonitor,
+  showDescription,
+  midiLearnAutoAdvance,
+  pasteAutoAdvance,
+  displayRowIndexAsHex,
+  initDisplayPreferences,
+  watchDisplayRowIndexAsHex
+} = usePanelState();
 
 // Track which row should auto-start MIDI learn
 const midiLearnAutoStartRow = ref<number | null>(null);
 
 // Track last learned MIDI message globally to prevent duplicates
 const lastLearnedMidiMessage = ref<string | null>(null);
-
-// Watch and persist panel visibility
-watch(showSelectionToolbar, (val) => {
-  localStorage.setItem(SELECTION_TOOLBAR_VISIBLE_KEY, val.toString());
-  if (!val) selectionToolbarExpanded.value = false;
-});
-watch(showMidiMonitor, (val) => {
-  localStorage.setItem(MIDI_MONITOR_VISIBLE_KEY, val.toString());
-  if (!val) midiMonitorExpanded.value = false;
-});
-watch(showVariableMonitor, (val) => {
-  localStorage.setItem(VARIABLE_MONITOR_VISIBLE_KEY, val.toString());
-  if (!val) variableMonitorExpanded.value = false;
-});
-watch(showLogMonitor, (val) => {
-  localStorage.setItem(LOG_MONITOR_VISIBLE_KEY, val.toString());
-  if (!val) logMonitorExpanded.value = false;
-});
-watch(showDescription, (val) => {
-  localStorage.setItem(DESCRIPTION_VISIBLE_KEY, val.toString());
-  if (!val) globalDocPanelExpanded.value = false;
-});
-watch(midiLearnAutoAdvance, (val) => {
-  localStorage.setItem(MIDI_LEARN_AUTO_ADVANCE_KEY, val.toString());
-});
-
-watch(pasteAutoAdvance, (val) => {
-  localStorage.setItem(PASTE_AUTO_ADVANCE_KEY, val);
-});
-
-// Row index display format (hex/decimal)
-const displayRowIndexAsHex = ref(false);
-const ROW_INDEX_DISPLAY_KEY = 'row-index-display-hex';
 
 const { isSupported: midiSupported } = useMidi();
 
@@ -798,15 +735,11 @@ function handleKeyDown(event: KeyboardEvent): void {
 onMounted(() => {
   document.addEventListener('keydown', handleKeyDown);
   // Load row index display preference from localStorage
-  displayRowIndexAsHex.value = localStorage.getItem(ROW_INDEX_DISPLAY_KEY) === 'true';
+  initDisplayPreferences();
 });
 
-// Persist row index display preference
-watch(displayRowIndexAsHex, (val) => {
-  localStorage.setItem(ROW_INDEX_DISPLAY_KEY, val.toString());
-  // Re-run analysis to update row indices in warning messages
-  analyzeDocument();
-});
+// Re-run analysis when row index display format changes
+watchDisplayRowIndexAsHex(() => analyzeDocument());
 
 // Auto-save to cache when document changes
 watch(
